project/requirements.txt
Flask==3.0.3
Flask-Login==0.6.3
Werkzeug==3.0.4
python-dotenv==1.0.1
psycopg2-binary==2.9.9
